Move Part type constants to directive.ts
